requirements.txt
opencv-python
torch
torchvision
transformers
pillow
numpy
scikit-learn
pandas
matplotlib
